fixing catalog UI, CRUD validation and more

+++ database/migrations/2022_12_02_074855_create_products_table.php
@@ -24,7 +24,7 @@ class CreateProductsTable extends Migration
             $table->text('src_img')->nullable();
             $table->integer('category_id');
             $table->integer('stock')->nullable();
-            $table->float('harga')->nullable();
+            $table->bigInteger('harga')->nullable();
             $table->text('description')->nullable();
             $table->timestamps();
         });
